Fix location search error by improving fallback behavior
- SearchService now falls back to MapKit when Nominatim returns
  server errors (non-2xx responses), not just network failures
- Changed placeId type from Int to Int64 to handle large IDs from
  the Nominatim API more safely
- Only throws error when both Nominatim and MapKit search providers fail

+++ ios/RoadRank/Models/SearchResult.swift
@@ -3,7 +3,7 @@ import MapKit
 
 // MARK: - Nominatim Search Result
 struct NominatimResult: Codable, Identifiable {
-    let placeId: Int
+    let placeId: Int64
     let lat: String
     let lon: String
     let displayName: String
@@ -17,7 +17,7 @@ struct NominatimResult: Codable, Identifiable {
         case type, importance
     }
 
-    var id: Int { placeId }
+    var id: Int64 { placeId }
 
     var coordinate: CLLocationCoordinate2D {
         CLLocationCoordinate2D(
